ubuntu-core-rootfs: deal with (broken) symlink in sync_dirs
The sync_dirs() has a test -e that fails to take broken symlinks
into account. This means that symlinks that are broken when
sync_dir runs (which is a very special environment) are not
copied. This breaks e.g. the symlink snapd.snap-repair.timer
in /etc/systemd/system/timer-target.wants.