move most of the customization into the core snap build
The customization of the core snap is currently divided into two
parts: ubuntu-core-config (the debian package) and the core snap
snapcraft. This leads to confusion like where the default netplan
configuration is generated. Merging most of ubuntu-core-config
into the snap build should fix this.

The coresponding core snap build PR is here:
https://github.com/snapcore/core/pull/83