config/etc/system-image/writable-paths: do not transition /snap
Newly build snaps have a "/snap" directory that contains some
metadata about the snap like "manifest.yaml" and "snapcraft.yaml".

The current writable-paths config sets /snap to transitional which
means that any files on the core snap under /snap will be copied
to the writable space. This is undesired, we just need an empty
/snap that is writable. This PR implements this.