scripts: add fixrtc-munt script to fix time issues on pi2,3
The fixrtc script we currently use to work around the issue that
the system clock in incorrect on a system without battery RTC
is not working well because the last mount time is moving
very slowly on xenial systems, see the forum discussion in
https://forum.snapcraft.io/t/1268/9 for details.

This version of fixrtc-mount will look at the "snaps" directory
to get a more accurate estimate of the system time.

Thanks to Alfonso Sanchez-Beato for the script.