many: drop static files for "core" snap from the package
The static files that used to be delivered via this package
are now part of the core snapcraft build via:
https://github.com/snapcore/core/pull/83

This makes the files unnecessary and they can be removed.

Moving the files into the snapcraft build removes one
level of indirection where we have to build the package,
upload to the ppa:snappy-dev/image PPA, wait for it to build
and then build core. With the PR#83 all these steps are
unnecessary now.